Migrated solution to .NET 6 framework.
