+++ samples/server/aspnet/appsettings.json
@@ -10,8 +10,8 @@
   "AllowedHosts": "*",
 
   // Our configuration
-  "host": "https://sample-commandcentre-server:8904",
-  "apiKey": "02FB-970F-0898-6BCB-D562-4BA3-C471-BBF3",
+  "host": "https://localhost:8904",
+  "apiKey": "E7CC-EB77-8F4A-4EA9-6452-B7F8-B84E-EAB1",
 
   "clientCertificatePfx": "clientcertificate.pfx",
   "clientCertificatePfxPassword": "secret",
